updated user login page

+++ projects/Javascript Projects/User Login/script.js
@@ -1,27 +1,24 @@
+let users = JSON.parse(localStorage.getItem('users')) || {} ;
 
 class User {
     
     constructor(firstname, lastname, email, username, password){
-        let id = parseInt(localStorage.getItem('ID')) || 1;
         this.firstname = firstname;
         this.lastname = lastname;
         this.email = email;
         this.username = username;
         this.password = password;
-        this.id = id;
         this.loggedIn = "Disconnected";
 
-        localStorage.setItem(`${id}`, JSON.stringify(this));
-        localStorage.setItem('ID', id + 1);
+        users[username] = this;
+        localStorage.setItem(`users`, JSON.stringify(users));
     }
 }
 
 const populateTable = () =>{
-    const numOfUsers = parseInt(localStorage.getItem('ID')) - 1;
-    for (i = 1; i <= numOfUsers; i++){
-        let user = JSON.parse(localStorage.getItem(i));
-        addUserToTable(user);
-    }
+    for (let username in users){
+        addUserToTable(users[username])
+    }  
 }
 
 document.addEventListener('DOMContentLoaded', ()=>{
@@ -32,26 +29,32 @@ const createUser = () => {
     let firstName = document.querySelector('#firstname').value;
     let lastName = document.querySelector('#lastname').value;
     let email = document.querySelector('#email').value;
-    let username = document.querySelector('#regUsername').value;
+    let newUsername = document.querySelector('#regUsername').value;
     let password = document.querySelector('#regPassword').value;
 
-    let newUser = new User(firstName, lastName, email, username, password);
-    
+    for (let username in users){
+        if (username === newUsername){
+            alert("that username is already in use")
+            return
+        } 
+    } 
+
+    let newUser = new User(firstName, lastName, email, newUsername, password);
     addUserToTable(newUser);
 }
 
 const addUserToTable = (user) => {
         const newRow = `
-            <tr>
+            <tr data-username="${user.username}">
                 <td>${user.firstname}</td>
                 <td>${user.lastname}</td>
                 <td>${user.email}</td>
-                <td>${user.username}</td>
+                <td class="usernameCol">${user.username}</td>
                 <td>${user.password}</td>
                 <td class="login-state">${user.loggedIn}</td>
-                <td><button class="form-button" id="logout-button" data-id="${user.id}">Log Out</button></td>
-                <td><button class="form-button" data-id="${user.id}">Edit</button></td>
-                <td><button class="form-button" data-id="${user.id}">Delete</button></td>
+                <td><button class="form-button" id="logout-button">Log Out</button></td>
+                <td><button class="form-button">Edit</button></td>
+                <td><button class="form-button">Delete</button></td>
             </tr>
         `;
 
@@ -64,19 +67,25 @@ document.querySelector('#register').addEventListener('click', (e) => {
 });
 
 const logIn = () => {
-    let username = document.querySelector('#loginUsername').value;
-    let password = document.querySelector('#loginPassword').value;
+    let loginUsername = document.querySelector('#loginUsername').value;
+    let loginPassword = document.querySelector('#loginPassword').value;
 
-    const numOfUsers = parseInt(localStorage.getItem('ID')) - 1;
-    for (i = 1; i <= numOfUsers; i++){
-        let user = JSON.parse(localStorage.getItem(i));
-        if (user.username === username && user.password === password){
-            user.loggedIn = "Connected";
-            loginState = "Connected"
-            localStorage.setItem(user.id, JSON.stringify(user));
-            document.querySelector(".login-state").innerHTML = "Connected"
+    for (let username in users){
+        if (username === loginUsername && users[username].password === loginPassword){
+            if (users[username].loggedIn === "Connected"){
+                alert('you are already logged in')
+                return
+            }
+            users[username].loggedIn = "Connected"
+            localStorage.setItem(`users`, JSON.stringify(users))
+            const userRow = document.querySelector(`tr[data-username="${loginUsername}"]`);
+            userRow.querySelector('.login-state').innerHTML = "Connected";
+            return
         }
     }
+
+    alert('the username or password is incorrect')
+    
 }
 
 document.querySelector('#login').addEventListener(('click'), (e) =>{
@@ -85,35 +94,31 @@ document.querySelector('#login').addEventListener(('click'), (e) =>{
 });
 
 const logOut = (e) =>{
-    const userId = e.target.dataset.id;
-    const numOfUsers = parseInt(localStorage.getItem('ID')) - 1;
-    for (i = 1; i <= numOfUsers; i++){
-        let user = JSON.parse(localStorage.getItem(i));
-        if (parseInt(userId) === user.id){
-            user.loggedIn = "Disconnected";
-            localStorage.setItem(user.id, JSON.stringify(user));
-            document.querySelector('.login-state').innerHTML = "Disconnected";
-        }
-    }
+    const userId = e.target.closest('tr');
+    const logoutUser = userId.querySelector('.usernameCol').innerHTML;
+    users[logoutUser].loggedIn = "Disconnected";
+    userId.querySelector('.login-state').innerHTML = "Disconnected";
+    localStorage.setItem(`users`, JSON.stringify(users))
 }
 
 const deleteUser = (e) => {
-    const userId = e.target.dataset.id;
-    const numOfUsers = parseInt(localStorage.getItem('ID')) - 1;
-    for (i = 1; i <= numOfUsers; i++){
-        let user = JSON.parse(localStorage.getItem(i));
-        if (parseInt(userId) === user.id){
-            localStorage.removeItem(user.id);
-            populateTable();
-        }
-    }
-    console.log(e);
-    
+    const userId = e.target.closest('tr');
+    userId.remove();
+
+    const deleteUser = userId.querySelector('.usernameCol').innerHTML;
+    delete users[deleteUser];
+    localStorage.setItem(`users`, JSON.stringify(users))
 }
 
 document.querySelector('#userTable').addEventListener(('click'), (e) =>{
-    logOut(e);
-    deleteUser(e);
+    if (e.target.innerHTML === "Log Out"){
+        logOut(e);
+    }
+    
+    if (e.target.innerHTML === "Delete"){
+        deleteUser(e)
+    }
+    
 });
